Test compression with invalid content type for all webservers
See gh-18337
